refactor(components): moves Dice to shared components
